moved register web components into shared folder for use in all modules. added the skeleton for the callout card
