+++ projects/buttons/src/lib/buttons.module.ts
@@ -4,19 +4,22 @@ import { createCustomElement } from '@angular/elements';
 
 import { ButtonComponent } from './components/button/button.component';
 import { ButtonGroupComponent } from './components/button-group/button-group.component';
+import { ElementDef } from 'projects/shared/interfaces';
+import { RegisterWebComponents } from 'projects/shared/helpers';
+
+const angularElements: ElementDef[] = [
+  { key: 'cns-button', component: ButtonComponent },
+  { key: 'cns-button-group', component: ButtonGroupComponent }
+];
 
 @NgModule({
-  declarations: [ButtonComponent, ButtonGroupComponent],
+  declarations: [angularElements.map(elementDef => elementDef.component)],
   imports: [CommonModule],
-  entryComponents: [ButtonComponent, ButtonGroupComponent]
+  entryComponents: [angularElements.map(elementDef => elementDef.component)]
 })
 export class ButtonsModule {
   constructor(private injector: Injector) {
-    const buttonElement = createCustomElement(ButtonComponent, { injector });
-    const buttonGroupElement = createCustomElement(ButtonGroupComponent, { injector });
-
-    customElements.define('cns-button', buttonElement);
-    customElements.define('cns-button-group', buttonGroupElement);
+    RegisterWebComponents(angularElements);
   }
 
   ngDoBootstrap() {}
