+++ projects/layouts/src/lib/components/callout-card/callout-card.component.html
