+++ projects/shared/helpers/index.ts
@@ -1,0 +1,1 @@
+export * from './register-webcomponents';
