+++ projects/shared/helpers/register-webcomponents.ts
@@ -1,0 +1,13 @@
+import { ElementDef } from '../interfaces';
+import { createCustomElement } from '@angular/elements';
+
+const RegisterWebComponents = (angularElements: ElementDef[]) => {
+  angularElements.forEach(element => {
+    const newElement = createCustomElement(element.component, { injector });
+    customElements.define(element.key, newElement);
+  });
+};
+
+export {
+  RegisterWebComponents
+};
