+++ projects/shared/interfaces/elementdef.interface.ts
@@ -1,0 +1,8 @@
+interface ElementDef {
+  key: string;
+  component: any;
+}
+
+export {
+  ElementDef
+};
