+++ projects/titles-and-text/src/lib/titles-and-text.module.ts
@@ -6,23 +6,24 @@ import { DisplayComponent } from './components/display/display.component';
 import { TitleComponent } from './components/title/title.component';
 import { TextComponent } from './components/text/text.component';
 import { CaptionComponent } from './components/caption/caption.component';
+import { ElementDef } from 'projects/shared/interfaces';
+import { RegisterWebComponents } from 'projects/shared/helpers';
+
+const angularElements: ElementDef[] = [
+  { key: 'cns-display', component: DisplayComponent },
+  { key: 'cns-title', component: TitleComponent },
+  { key: 'cns-text', component: TextComponent },
+  { key: 'cns-caption', component: CaptionComponent }
+];
 
 @NgModule({
-  declarations: [DisplayComponent, TitleComponent, TextComponent, CaptionComponent],
+  declarations: [angularElements.map(element => element.component)],
   imports: [CommonModule],
-  entryComponents: [DisplayComponent, TitleComponent, TextComponent, CaptionComponent]
+  entryComponents: [angularElements.map(element => element.component)]
 })
 export class TitlesAndTextModule {
   constructor(private injector: Injector) {
-    const displayElement = createCustomElement(DisplayComponent, { injector });
-    const titleElement = createCustomElement(TitleComponent, { injector });
-    const textElement = createCustomElement(TextComponent, { injector });
-    const captionElement = createCustomElement(CaptionComponent, { injector });
-
-    customElements.define('cns-display', displayElement);
-    customElements.define('cns-title', titleElement);
-    customElements.define('cns-text', textElement);
-    customElements.define('cns-caption', captionElement);
+    RegisterWebComponents(angularElements);
   }
 
   ngDoBootstrap() {}
